Create errors actions and reducer
