+++ client/web/actions/session_actions.js
@@ -1,9 +1,8 @@
 import * as SessionAPIUtil from '../util/session_api_util';
+import * as ErrorActions from './errors_actions';
 
 export const START_SESSION = 'START_SESSION';
 export const END_SESSION = 'END_SESSION';
-export const RECEIVE_ERRORS = 'RECEIVE_ERRORS';
-export const CLEAR_ERRORS = 'CLEAR_ERRORS';
 
 export const startSession = res => ({
   type: START_SESSION,
@@ -15,15 +14,6 @@ const quitSession = res => ({
   res
 });
 
-export const receiveErrors = errs => ({
-  type: RECEIVE_ERRORS,
-  errs
-});
-
-const clearErrors = () => ({
-  type: CLEAR_ERRORS
-});
-
 export const fetchCurrentSession = () => dispatch =>
   SessionAPIUtil.fetchCurrentSession()
     .then(res => dispatch(startSession(res)), err => console.log(err, 'error'))
@@ -32,7 +22,7 @@ export const fetchCurrentSession = () => dispatch =>
 export const login = user => dispatch =>
   SessionAPIUtil.startSession(user)
     .then(res => dispatch(startSession(res)), err => (
-      dispatch(receiveErrors(err.response.data))
+      dispatch(ErrorActions.receiveErrors(err.response.data))
     ))
     .catch(err => console.log(err));
 
@@ -40,5 +30,3 @@ export const logout = () => dispatch =>
   SessionAPIUtil.endSession()
     .then(res => dispatch(quitSession(res)))
     .catch(err => console.log(err));
-
-export const removeErrors = () => dispatch => dispatch(clearErrors);
