+++ client/web/actions/user_actions.js
@@ -1,5 +1,6 @@
 import * as UserAPIUtil from '../util/user_api_util';
-import { startSession, receiveErrors } from './session_actions';
+import * as ErrorActions from './errors_actions';
+import { startSession } from './session_actions';
 
 export const RECEIVE_USER = 'RECEIVE_USER';
 
@@ -12,6 +13,6 @@ export const createUser = newUser => dispatch =>
   UserAPIUtil.createUser(newUser)
     .then(
       res => dispatch(startSession(res)),
-      err => dispatch(receiveErrors(err.response.data))
+      err => dispatch(ErrorActions.receiveErrors(err.response.data))
     )
     .catch(err => console.log(err));
