+++ client/web/reducers/error/errors_reducer.js
@@ -1,0 +1,16 @@
+import {
+  RECEIVE_ERRORS,
+  CLEAR_ERRORS
+} from '../../actions/errors_actions';
+
+export default (state = [], action) => {
+  Object.freeze(state);
+  switch (action.type) {
+    case RECEIVE_ERRORS:
+      return action.errs;
+    case CLEAR_ERRORS:
+      return [];
+    default:
+      return state;
+  }
+};
